+++ src/OpenTelemetry.AutoInstrumentation/Instrumentation.cs
@@ -369,9 +369,9 @@ internal static class Instrumentation
                     break;
                 case TracerInstrumentation.Kafka:
                     break;
-                case TracerInstrumentation.RabbitMq:
-                    break;
                 case TracerInstrumentation.OracleMda:
+                    break;
+                case TracerInstrumentation.RabbitMq:
                     break;
                 default:
                     Logger.Warning($"Configured trace instrumentation type is not supported: {instrumentation}");