+++ CHANGELOG.md
@@ -5,9 +5,27 @@ All notable changes to this component are documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 This component adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased](https://github.com/open-telemetry/opentelemetry-dotnet-instrumentation/compare/v1.8.0..HEAD)
+## [Unreleased](https://github.com/open-telemetry/opentelemetry-dotnet-instrumentation/compare/v1.9.0..HEAD)
 
 ### Added
+
+- Support for [RabbitMQ.Client](https://www.nuget.org/packages/RabbitMQ.Client/)
+  traces instrumentation for versions `7.0.0`+.
+
+### Changed
+
+#### Dependency updates
+
+### Deprecated
+
+### Removed
+
+- Support for macOS Monterey 12 x64.
+  macOs libraries are built and tested against [macOS Ventura 13 x64](https://github.com/actions/runner-images/blob/main/images/macos/macos-13-Readme.md).
+
+### Fixed
+
+## [1.9.0](https://github.com/open-telemetry/opentelemetry-dotnet-instrumentation/releases/tag/v1.9.0)
 
 ### Changed
 
@@ -15,12 +33,20 @@ This component adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.h
 
 #### Dependency updates
 
+- Following packages updated
+  - `OpenTelemetry.Instrumentation.Wcf` from `1.0.0-rc.17` to `1.0.0-rc.18`,
 - .NET Framework only, following packages updated
+  - `Google.Protobuf` updated from `3.28.2` to `3.28.3`,
   - `Microsoft.Extensions.DependencyInjection` from `8.0.0` to `8.0.1`,
   - `Microsoft.Extensions.DependencyInjection.Abstractions` from `8.0.1` to `8.0.2`,
+  - `System.Diagnostics.DiagnosticSource` from `8.0.0` to `8.0.1`,
   - `System.Text.Json` from `8.0.4` to `8.0.5`.
 
 ### Deprecated
+
+- Support for macOS Monterey 12 x64.
+  All further releases will be supporting [macOS Ventura 13 x64](https://github.com/actions/runner-images/blob/main/images/macos/macos-13-Readme.md)
+  and newer.
 
 ### Removed
 
@@ -32,10 +58,6 @@ This component adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.h
   `OTEL_TRACES_EXPORTER`
   `OTEL_METRICS_EXPORTER`
   `OTEL_LOGS_EXPORTER`.
-- Support for macOS Monterey 12 x64.
-  macOs libraries are built and tested against [macOS Ventura 13 x64](https://github.com/actions/runner-images/blob/main/images/macos/macos-13-Readme.md).
-
-### Fixed
 
 ## [1.8.0](https://github.com/open-telemetry/opentelemetry-dotnet-instrumentation/releases/tag/v1.8.0)